requirements.txt
asgiref==3.6.0
Django==4.1.5
pytz==2022.7.1
sqlparse==0.4.3
djangorestframework==3.12.4
django-cors-headers==3.10.0
drf_yasg==1.20.0
pillow
django-filter
djangorestframework-simplejwt
django-cors-headers
datetime
pyparsing 
pydot
django-extensions